+++ test/rails_app/db/migrate/20200411171322_create_rodauth.rb
@@ -10,10 +10,12 @@ class CreateRodauth < superclass
 
   def change
     create_table :accounts do |t|
-      t.string :email, null: false, index: { unique: true }
+      t.string :email, null: false
       t.string :status, null: false, default: "unverified"
       t.string :password_hash
     end
+
+    add_index :accounts, :email, unique: true, where: "status IN ('unverified', 'verified')"
 
     # Used if storing password hashes in a separate table (default)
     create_table :account_password_hashes do |t|